The first and second part of the data analysis
This is the first part and second part of my work, the events detection and evaluation of the driving behaviors. It includes four bagfile analysis function (which records some basic strategies to process the bagfile using bagpy) and 10 events detection metrics, of which 4 of them are defined referred to existed literature, and 6 of them are original defined by my self. I have written all the intsructions and input, output, comment below, hopefully, very clear. For more information, please refer to my work report.