The third part of data analysis
This is the third part of my work. For this part, we should change the parameters of the ROS simulation, then get the command velocity from simulation results and make comparisons with the two. By checking the difference, we may further define new metrics to evaluate the performance of the controllers.